Refactor: Move creator components to features directory - Moved CreatorCard to components/features/creator/card - Moved CreatorMedia to components/features/creator/media - Moved CreatorRating to components/features/creator/rating - Moved PortfolioPreview to components/features/creator/portfolio - Moved creator types to components/features/creator/types - Created READMEs for each directory explaining purpose - Maintained backward compatibility in original files
