+++ COMPONENT-MOVEMENT-PLAN.md
@@ -42,10 +42,17 @@ This document outlines the plan for moving components to the new directory struc
 
 #### Creator Components
 
-| Component | Source Path | Destination Path | Import Updates | Testing Criteria |
-|-----------|-------------|-----------------|----------------|------------------|
-| CreatorCard.tsx | /src/components/creator/CreatorCard.tsx | /src/components/features/creator/card/CreatorCard.tsx | Update imports from `@/components/creator/CreatorCard` to `@/components/features/creator/card/CreatorCard` | Verify creator cards render correctly throughout the site |
-| ... | ... | ... | ... | ... |
+| Component | Source Path | Destination Path | Status | Testing Criteria |
+|-----------|-------------|-----------------|--------|------------------|
+| CreatorCard.tsx | /src/components/creator/CreatorCard.tsx | /src/components/features/creator/card/CreatorCard.tsx | ✅ Completed | Verify creator cards render correctly throughout the site |
+| CreatorMedia.tsx | /src/components/creator/CreatorMedia.tsx | /src/components/features/creator/media/CreatorMedia.tsx | ✅ Completed | Verify creator media displays correctly in cards |
+| CreatorRating.tsx | /src/components/creator/CreatorRating.tsx | /src/components/features/creator/rating/CreatorRating.tsx | ✅ Completed | Verify ratings display correctly in creator cards |
+| PortfolioPreview.tsx | /src/components/creator/PortfolioPreview.tsx | /src/components/features/creator/portfolio/PortfolioPreview.tsx | ✅ Completed | Verify portfolio previews display in creator cards |
+| types.ts | /src/components/creator/types.ts | /src/components/features/creator/types/index.ts | ✅ Completed | Verify types are correctly exported and used |
+| CreatorInfo.tsx | /src/components/creator/CreatorInfo.tsx | /src/components/features/creator/info/CreatorInfo.tsx | Not Started | Verify creator info displays correctly |
+| NotableClients.tsx | /src/components/creator/NotableClients.tsx | /src/components/features/creator/clients/NotableClients.tsx | Not Started | Verify notable clients section displays correctly |
+| CreatorTags.tsx | /src/components/creator/CreatorTags.tsx | /src/components/features/creator/tags/CreatorTags.tsx | Not Started | Verify creator tags display correctly |
+| PortfolioGallery.tsx | /src/components/creator/PortfolioGallery.tsx | /src/components/features/creator/portfolio/PortfolioGallery.tsx | Not Started | Verify portfolio gallery displays correctly |
 
 ## Implementation Notes
