+++ src/components/features/creator/card/index.ts
@@ -1,0 +1,1 @@
+export * from './CreatorCard'
