+++ src/components/features/creator/media/README.md
@@ -1,0 +1,14 @@
+# Creator Media Components
+
+This directory contains components for rendering creator media (images, videos) in the application.
+
+## Components
+
+- `CreatorMedia.tsx` - Renders creator images/videos with performance optimizations for mobile vs desktop
+- `index.ts` - Barrel file for clean exports
+
+## Usage
+
+```tsx
+import { CreatorMedia } from '@/components/features/creator';
+```
