+++ src/components/features/creator/media/index.ts
@@ -1,0 +1,1 @@
+export { CreatorMedia, getMedia } from './CreatorMedia';
