+++ src/components/features/creator/portfolio/README.md
@@ -1,0 +1,14 @@
+# Creator Portfolio Components
+
+Components for displaying creator portfolio samples and related gallery views.
+
+## Components
+
+- `PortfolioPreview.tsx` - Grid display of portfolio preview images
+- `index.ts` - Barrel file for clean exports
+
+## Usage
+
+```tsx
+import { PortfolioPreview } from '@/components/features/creator';
+```
