+++ src/components/features/creator/portfolio/index.ts
@@ -1,0 +1,1 @@
+export { PortfolioPreview } from './PortfolioPreview';
