+++ src/components/features/creator/rating/README.md
@@ -1,0 +1,14 @@
+# Creator Rating Components
+
+Components related to displaying creator ratings and availability status.
+
+## Components
+
+- `CreatorRating.tsx` - Displays star rating, review count, and availability status
+- `index.ts` - Barrel file for clean exports
+
+## Usage
+
+```tsx
+import { CreatorRating } from '@/components/features/creator';
+```
