+++ src/components/features/creator/rating/index.ts
@@ -1,0 +1,1 @@
+export { CreatorRating } from './CreatorRating';
