+++ src/components/features/creator/types/README.md
@@ -1,0 +1,16 @@
+# Creator Type Definitions
+
+Type definitions for creator-related data structures used throughout the application.
+
+## Types
+
+- `Creator` - Main creator data interface
+- `CreatorCardProps` - Props for the CreatorCard component
+- `AvailabilityStatus` - Availability status options
+- `NotableClient` - Interface for creator's notable clients
+
+## Usage
+
+```tsx
+import { Creator, AvailabilityStatus } from '@/components/features/creator';
+```
